Rename clients classes
