+++ src/index.ts
@@ -1,7 +1,7 @@
 import * as querystring from 'querystring';
-import { ClientDeviceTracker } from './client/ClientDeviceTracker';
+import { DeviceTrackerClient } from './client/DeviceTrackerClient';
 import { ScrcpyClient, StreamParams } from './client/ScrcpyClient';
-import { ShellParams, ClientShell } from './client/ClientShell';
+import { ShellParams, ShellClient } from './client/ShellClient';
 
 window.onload = function (): void {
     const hash = location.hash.replace(/^#!/, '');
@@ -9,9 +9,9 @@ window.onload = function (): void {
     const action = parsedQuery.action;
     if (action === ScrcpyClient.ACTION && typeof parsedQuery.udid === 'string') {
         ScrcpyClient.start(parsedQuery as StreamParams);
-    } else if (action === ClientShell.ACTION && typeof parsedQuery.udid === 'string') {
-        ClientShell.start(parsedQuery as ShellParams);
+    } else if (action === ShellClient.ACTION && typeof parsedQuery.udid === 'string') {
+        ShellClient.start(parsedQuery as ShellParams);
     } else {
-        ClientDeviceTracker.start();
+        DeviceTrackerClient.start();
     }
 };
